Integrate stock test into new update product suite relying on unified command and independe combination stock command
